+++ core/server/api_container/server/startosis_engine/startosis_interpreter.go
@@ -41,7 +41,6 @@ const (
 	multipleInterpretationErrorMsg = "Multiple errors caught interpreting the Starlark script. Listing each of them below."
 
 	skipImportInstructionInStacktraceValue = "import_module"
-
 )
 
 type StartosisInterpreter struct {
@@ -144,7 +143,7 @@ func (interpreter *StartosisInterpreter) buildBindings(thread *starlark.Thread, 
 		// Kurtosis instructions - will push instructions to the queue that will affect the enclave state at execution
 		add_service.AddServiceBuiltinName:                starlark.NewBuiltin(add_service.AddServiceBuiltinName, add_service.GenerateAddServiceBuiltin(instructionsQueue, interpreter.serviceNetwork, interpreter.factsEngine)),
 		assert.AssertBuiltinName:                         starlark.NewBuiltin(assert.AssertBuiltinName, assert.GenerateAssertBuiltin(instructionsQueue, interpreter.recipeExecutor, interpreter.serviceNetwork)),
-		exec.ExecBuiltinName:                             starlark.NewBuiltin(exec.ExecBuiltinName, exec.GenerateExecBuiltin(instructionsQueue, interpreter.serviceNetwork)),
+		exec.ExecBuiltinName:                             starlark.NewBuiltin(exec.ExecBuiltinName, exec.GenerateExecBuiltin(instructionsQueue, interpreter.serviceNetwork, interpreter.recipeExecutor)),
 		extract.ExtractBuiltinName:                       starlark.NewBuiltin(extract.ExtractBuiltinName, extract.GenerateExtractInstructionBuiltin(instructionsQueue, interpreter.recipeExecutor, interpreter.serviceNetwork)),
 		get_value.GetValueBuiltinName:                    starlark.NewBuiltin(get_value.GetValueBuiltinName, get_value.GenerateGetValueBuiltin(instructionsQueue, interpreter.recipeExecutor, interpreter.serviceNetwork)),
 		kurtosis_print.PrintBuiltinName:                  starlark.NewBuiltin(kurtosis_print.PrintBuiltinName, kurtosis_print.GeneratePrintBuiltin(instructionsQueue, interpreter.recipeExecutor, interpreter.serviceNetwork)),
@@ -234,4 +233,3 @@ func generateInterpretationError(err error) *startosis_errors.InterpretationErro
 	}
 	return startosis_errors.NewInterpretationError("UnknownError: %s\n", err.Error())
 }
-